extends the idea pushed by @jbrockmendel in #43285 

```python
>>> df = pd.DataFrame(np.random.randn(10000, 1))
>>> % timeit df.style.applymap_index(lambda v: "color: blue;")._compute()
29.7 ms ± 396 µs per loop (mean ± std. dev. of 7 runs, 10 loops each) # Master
24.4 ms ± 347 µs per loop (mean ± std. dev. of 7 runs, 10 loops each) # PR
```